SHAR-2 Implement a basic front end for uploading a video
We now have a basic front end for a user to select and upload a video.

We also show the user the remaining mb and percentage of the video to be
uploaded during the upload.

Currently, we only store the video in memory, but in the future we will
store the raw video into a NoSQL database to be encoded.